# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  common.BaseResponse-dto_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/dto.LoginResponse'
      message:
        type: string
      status:
        type: integer
    type: object
  common.BaseResponse-dto_RefreshResponse:
    properties:
      data:
        $ref: '#/definitions/dto.RefreshResponse'
      message:
        type: string
      status:
        type: integer
    type: object
  common.BaseResponse-dto_UserDTO:
    properties:
      data:
        $ref: '#/definitions/dto.UserDTO'
      message:
        type: string
      status:
        type: integer
    type: object
  common.BaseResponse-string:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
  dto.CreateUser:
    properties:
      email:
        example: budipambudi@gmail.com
        type: string
      password:
        example: supersecretpassword
        type: string
      username:
        example: Budi Pambudi
        type: string
    type: object
...
  dto.LoginDTO:
    properties:
      email:
        example: fadelanumah@gmail.com
        type: string
      password:
        example: Secretpassword@123
        type: string
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        example: Grgws!23412sA
        type: string
      refresh_token:
        example: Grgws!23412sA
        type: string
    type: object
  dto.RefreshDTO:
    properties:
      refresh_token:
        type: string
    type: object
  dto.RefreshResponse:
    properties:
      access_token:
        type: string
    type: object
...
  dto.UserDTO:
    properties:
      created_at:
        example: "1617181723"
        type: string
      email:
        example: budipambudi@gmail.com
        type: string
      id:
        example: a53515e3-5a7f-440b-82f6-3d84ac7ce746
        type: string
      updated_at:
        example: "1617181723"
        type: string
      username:
        example: Budi Pambudi
        type: string
    type: object
info:
  contact: {}
  description: Go, Gin, and Postgre Base Project
  title: Go, Gin, and Postgre Base Project
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: request body for log in [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse-dto_LoginResponse'
      summary: Login User
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user
      parameters:
      - description: request body for logout user [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse-string'
      summary: Logout user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token
      parameters:
      - description: request body for refresh access token [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse-dto_RefreshResponse'
      summary: Refresh Access Token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register user
      parameters:
      - description: request body for register an user [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.BaseResponse-dto_UserDTO'
      summary: Register User
      tags:
      - auth
...
swagger: "2.0"